Some small TUI cleanup
